3_code/backend/requirements.txt
fastapi
uvicorn
sqlmodel
langgraph
langchain-core
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv
httpx
psycopg2-binary
bcrypt
google-generativeai>=0.8.3
stripe
google-auth
